chime: switch to QTI BT stack
In an attempt to fix some issues related to BT audio, we'll be switching
to the QTI Bluetooth stack rather than the AOSP one, testing is required
to see whether the issues are solved or not.

The audio configs may need to be replaced/modified to be fit for the new
BT stack.